requirements.txt
fastapi
uvicorn
pydantic
crewai
crewai-tools
langchain
langchain-openai
python-dotenv
python-multipart
setuptools
